refactored tests to try to get Travis to recognize them
